# contracts/openapi/domain-services.v1.yaml
openapi: 3.1.0
info:
  title: SurakshaSetu domain services
  version: 1.0.0
  description: |
    The tier contract between the orchestrator (Python) and the domain services (Java). It is the
    single source of truth: Java interfaces and models and Python models are generated from it.
    Model shapes follow the TDD (§3.5–§3.8) where it defines them.

    ## Conventions
    - **Money** is a decimal string of rupees (`^\d+(\.\d{1,2})?$`), never a JSON number.
    - **Dates** are ISO 8601 (`date` or `date-time`); date-times are UTC.
    - **Errors** are `application/problem+json` with `type`, `title`, `status`, `detail` and a
      machine-readable `code`.
    - **Auth**: a bearer token (a shared dev token locally; mTLS in production).
    - **Decisions** (`POST` eligibility, suitability, ranking and quotes): the request carries
      `pins.rules` and the service evaluates exactly that version, or answers 409
      `RULES_VERSION_UNKNOWN`. The response carries `decision_id`, the version it used
      (`rules_version`, `ranker_version` or `rating_version`), `inputs_sha256` and `reason_codes`.
    - Commission, margin, incentive, campaign and payout fields never appear in this contract.

    ## Problem codes
    | Status | code | When |
    | --- | --- | --- |
    | 404 | `NOT_FOUND` | Unknown UIN, consent id, notice, disclosure, pincode or occupation. |
    | 409 | `RULES_VERSION_UNKNOWN` | `pins.rules` names a version the service has not loaded. |
...

    ## Hashing
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/consent/notices/current:
    get:
      tags: [consent]
      operationId: getCurrentConsentNotice
      summary: The approved consent notice currently in force for a language.
      parameters:
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentNotice'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/consent/notices/{notice_version}:
    get:
      tags: [consent]
      operationId: getConsentNotice
      parameters:
        - name: notice_version
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentNotice'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/consent/records:
    post:
      tags: [consent]
      operationId: createConsentRecord
      summary: Record consent. Replaying the same Idempotency-Key returns the original record.
      description: >-
        `captured_at` is taken from the server clock; any client value is ignored. Reusing an
        Idempotency-Key with a different body is rejected with 409 `IDEMPOTENCY_KEY_REUSED`.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConsentRecordCreate'}
      responses:
        '200':
          description: A replay of the same Idempotency-Key; the original record.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentRecord'}
        '201':
          description: Created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentRecord'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/consent/records/{consent_id}:
    get:
      tags: [consent]
      operationId: getConsentRecord
      summary: A consent record, with valid_p1 recomputed as of the given instant.
      parameters:
        - $ref: '#/components/parameters/ConsentId'
        - $ref: '#/components/parameters/AsOf'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentRecord'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/consent/records/{consent_id}/purposes:
    post:
      tags: [consent]
      operationId: changeConsentPurpose
      summary: Grant or revoke one purpose. Appends to the purpose history.
      parameters:
        - $ref: '#/components/parameters/ConsentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PurposeGrant'}
      responses:
        '200':
          description: The record after the change.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentRecord'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/consent/records/{consent_id}/withdraw:
    post:
      tags: [consent]
      operationId: withdrawConsent
      summary: Withdraw consent (invariant I5).
      parameters:
        - $ref: '#/components/parameters/ConsentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConsentWithdrawal'}
      responses:
        '200':
          description: The record after withdrawal.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsentRecord'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Language:
      name: language
      in: query
      required: true
      schema: {type: string}
      example: en-IN
...
    AsOf:
      name: as_of
      in: query
      description: Evaluate as of this instant. Defaults to now.
      schema: {type: string, format: date-time}
...
    ConsentId:
      name: consent_id
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    Problem:
      description: Error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Problem:
      type: object
      required: [title, status, code]
      properties:
        type: {type: string, default: 'about:blank', description: 'RFC 9457: absent means about:blank.'}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance:
          type: string
          description: 'RFC 9457: the request path. It can carry ids or a pincode, so never log it.'
        code:
          type: string
          description: Machine-readable error code; the table in the API description lists them.
...
    Sha256Hex:
      type: string
      pattern: '^[0-9a-f]{64}$'
...
    ConsentNotice:
      type: object
      required: [notice_version, language, body, body_sha256, is_dummy]
      properties:
        notice_version: {type: string, examples: [2026.09.1-en]}
        language: {type: string}
        body: {type: string}
        body_sha256: {$ref: '#/components/schemas/Sha256Hex'}
        is_dummy: {type: boolean}
    PurposeId:
      type: string
      enum: [P1_NEEDS_RECO, P2_ADVISOR_CONTACT, P3_MARKETING]
    PurposeGrant:
      type: object
      required: [purpose_id, granted]
      properties:
        purpose_id: {$ref: '#/components/schemas/PurposeId'}
        granted: {type: boolean}
    ConsentMethod:
      type: string
      description: >-
        How consent was captured. TDD §3.5 says typed_affirmation; the DDL (§7.2) says
        parsed_affirmation, which is used everywhere.
      enum: [structured_action, parsed_affirmation]
    ConsentRecordCreate:
      type: object
      required: [session_id, subject_ref, notice_version, notice_sha256, language,
                 ai_disclosure_version, purposes, age_18_plus_declared, method]
      properties:
        session_id: {type: string, format: uuid}
        subject_ref: {type: string, format: uuid, description: Pseudonymous subject reference.}
        notice_version: {type: string}
        notice_sha256: {$ref: '#/components/schemas/Sha256Hex'}
        language: {type: string, description: The language the notice was shown in.}
        ai_disclosure_version: {type: string}
        purposes:
          type: array
          items: {$ref: '#/components/schemas/PurposeGrant'}
        age_18_plus_declared: {type: boolean}
        method: {$ref: '#/components/schemas/ConsentMethod'}
    ConsentRecord:
      type: object
      required: [consent_id, notice_version, notice_sha256, notice_language, ai_disclosure_version,
                 purposes, age_18_plus_declared, method, captured_at, valid_p1, valid_reasons]
      properties:
        consent_id: {type: string, format: uuid}
        notice_version: {type: string}
        notice_sha256: {$ref: '#/components/schemas/Sha256Hex'}
        notice_language: {type: string}
        ai_disclosure_version: {type: string}
        purposes:
          type: array
          description: Current state, one grant per purpose. The history stays in the domain tier.
          items: {$ref: '#/components/schemas/PurposeGrant'}
        age_18_plus_declared: {type: boolean}
        method: {$ref: '#/components/schemas/ConsentMethod'}
        captured_at: {type: string, format: date-time, description: 'Server clock, UTC.'}
        withdrawn_at: {type: [string, 'null'], format: date-time}
        valid_p1:
          type: boolean
          description: >-
            P1 granted, not withdrawn, 18+ declared, notice not superseded and within the consent
            TTL, as of the request's as_of.
        valid_reasons:
          type: array
          description: >-
            Why valid_p1 is false; empty when it is true. Codes: P1_NOT_GRANTED, WITHDRAWN,
            AGE_NOT_DECLARED, NOTICE_SUPERSEDED (a newer notice is in force for the language) and
            CONSENT_EXPIRED (older than the consent TTL). The last two send a resumed session back to S0.
          items: {type: string}
    ConsentWithdrawal:
      type: object
      properties:
        reason: {type: string}

    # --- Catalog (TDD §7.2, §7.3) -----------------------------------------------------------
...